- added props validation, added first map and css for component costume
