+++ src/App.jsx
@@ -1,7 +1,7 @@
 
 import { Header } from "./components/Header.jsx"
+import { MultiStepForm } from "./components/Multistepform.jsx";
 
-import { MultiStepForm } from "./components/SubminNew.jsx";
 
 export const App = () => {
     return (
