+++ src/components/Header.jsx
@@ -3,7 +3,7 @@ export const Header = () => {
         <header>
             <h1>Superpower Shenanigans</h1>
             <p>
-            Welcome to the "Superpower Shenanigans" survey! Get ready to create useless superpowers, choose funny scenarios, and design the silliest superhero costume. Let’s have some fun and discover your quirky superhero side!</p>
+                Welcome to the Superpower Shenanigans survey! Get ready to create useless superpowers, choose funny scenarios, and design the silliest superhero costume. Let’s have some fun and discover your quirky superhero side!</p>
         </header>
 
     )
