+++ src/components/costume/costume.css
@@ -1,0 +1,4 @@
+.form-container-costume {
+    background-color: green;
+    color: gold;
+}
